Update patch set 1
Patch Set 1:

I will try to look into this again tonight on fireball to see if it exists there. I'm not particularly fond of this hack. It would be worthwhile run dump() from IccCardProxy when this occurs to see if cardstate>2 only occurs when mRadioOn==false, in which case we could reassign the cardstate to CARDSTATE_ABSENT and would no longer be simply dumping the response.

Patch-set: 1
Label: Code-Review=0